fix(gateway): Resolve Rust 2024 edition compatibility issues
- Fix lifetime issues with borrowed values in async streams
- Update middleware to use http-body-util for BoxBody type
- Fix filter_map closures to use async move syntax
- Add explicit type annotations for impl Stream returns
- Fix BroadcastStreamRecvError pattern matching
- Add bytes crate dependency for body types

+++ services/gateway/src/services/metrics.rs
@@ -95,7 +95,7 @@ impl MetricsService {
                 country_name: row.get("country_name"),
                 requests: row.get::<i64, _>("requests") as u64,
                 bytes: row.get::<i64, _>("bytes") as u64,
-                unique_ips: row.get::<i64, _>("unique_ips") as u32,
+                unique_ips: row.get::<i64, _>("unique_ips") as u64,
                 blocked: false,
             })
             .collect();
